// packages/headless/src/builders/brick-builder/types.ts
export type FieldTypes =
	| "tab"
	| "text"
	| "wysiwyg"
	| "media"
	| "repeater"
	| "number"
	| "checkbox"
	| "select"
	| "textarea"
	| "json"
	| "colour"
	| "datetime"
	| "pagelink"
	| "link";

export enum FieldTypesEnum {
	Tab = "tab",
	Text = "text",
	Wysiwyg = "wysiwyg",
	Media = "media",
	Repeater = "repeater",
	Number = "number",
	Checkbox = "checkbox",
	Select = "select",
	Textarea = "textarea",
	JSON = "json",
	Colour = "colour",
	Datetime = "datetime",
	Pagelink = "pagelink",
	Link = "link",
}

// Custom Fields - fields define in the builder
export interface CustomField {
	type: FieldTypes;
	key: CustomFieldConfig["key"];
	title: CustomFieldConfig["title"];
	description?: CustomFieldConfig["description"];
	placeholder?: string;
	fields?: Array<CustomField>;
	default?: defaultFieldValues;

	presets?: string[];
	copy?: {
		true?: string;
		false?: string;
	};
	options?: Array<{
		label: string;
		value: string;
	}>;
	// Validation
	validation?: {
		zod?: z.ZodType<unknown>;
		required?: boolean;
		extensions?: string[];
		type?: "image" | "video" | "document" | "archive";
		maxGroups?: number;
		width?: {
			min?: number;
			max?: number;
		};
		height?: {
			min?: number;
			max?: number;
		};
	};
}

// ------------------------------------
// Validation types
export interface ValidationProps {
	type: FieldTypes;
	key: CustomFieldConfig["key"];
	value: unknown;
	referenceData?: MediaReferenceData | LinkReferenceData;
	flatFieldConfig: CustomField[];
}
export interface ValidationResponse {
	valid: boolean;
	message?: string;
}

export interface LinkReferenceData {
	target?: string | null;
	label?: string | null;
}

export interface MediaReferenceData {
	extension: string;
	width: number | null;
	height: number | null;
	type: string;
}

// ------------------------------------
// Custom Fields Config
export interface CustomFieldConfig {
	key: string;
	title?: string;
	description?: string;
}
// ...
export type defaultFieldValues =
	| TextConfig["default"]
	| WysiwygConfig["default"]
	| NumberConfig["default"]
	| CheckboxConfig["default"]
	| SelectConfig["default"]
	| TextareaConfig["default"]
	| JSONConfig["default"]
	| ColourConfig["default"]
	| DateTimeConfig["default"];